+++ CHANGES.rst
@@ -8,6 +8,7 @@ Changelog
   <https://github.com/mgedmin/check-python-versions/issues/22>`_).
 - Correctly detect the end of a YAML list even if it's not indented.
 - Try to preserve envlist order in ``tox.ini``.
+- Preserve quotes around Python versions in Travis matrix jobs.
 
 
 0.16.1 (2020-11-08)